Criado PersistirRelatorioPAPCommand
